+++ ios/RNKeyEvent.xcworkspace/contents.xcworkspacedata
@@ -1,0 +1,9 @@
+// !$*UTF8*$!
+<?xml version="1.0" encoding="UTF-8"?>
+<Workspace
+   version = "1.0">
+   <FileRef
+      location = "group:RNKeyEvent.xcodeproj">
+   </FileRef>
+</Workspace>
+
