+++ package.json
@@ -1,7 +1,6 @@
 {
   "name": "react-native-keyevent",
   "version": "0.1.2",
-  "main": "index.android.js",
   "description": "Capture external keyboard keys or remote control button events",
   "keywords": [
     "react-native",
